+++ apps/client/src/app/pages/portfolio/activities/create-or-update-activity-dialog/create-or-update-activity-dialog.component.ts
@@ -139,7 +139,8 @@ export class CreateOrUpdateActivityDialog implements OnDestroy {
       unitPriceInCustomCurrency: [
         this.data.activity?.unitPrice,
         Validators.required
-      ]
+      ],
+      updateAccountBalance: [false]
     });
 
     this.activityForm.valueChanges
@@ -411,7 +412,9 @@ export class CreateOrUpdateActivityDialog implements OnDestroy {
           : this.activityForm.controls['searchSymbol'].value.symbol,
       tags: this.activityForm.controls['tags'].value,
       type: this.activityForm.controls['type'].value,
-      unitPrice: this.activityForm.controls['unitPrice'].value
+      unitPrice: this.activityForm.controls['unitPrice'].value,
+      updateAccountBalance:
+        this.activityForm.controls['updateAccountBalance'].value
     };
 
     if (this.data.activity.id) {