+++ apps/client/src/app/services/import-activities.service.ts
@@ -59,7 +59,8 @@ export class ImportActivitiesService {
         quantity: this.parseQuantity({ content, index, item }),
         symbol: this.parseSymbol({ content, index, item }),
         type: this.parseType({ content, index, item }),
-        unitPrice: this.parseUnitPrice({ content, index, item })
+        unitPrice: this.parseUnitPrice({ content, index, item }),
+        updateAccountBalance: false
       });
     }
 
@@ -126,7 +127,8 @@ export class ImportActivitiesService {
     quantity,
     SymbolProfile,
     type,
-    unitPrice
+    unitPrice,
+    updateAccountBalance
   }: Activity): CreateOrderDto {
     return {
       accountId,
@@ -136,7 +138,8 @@ export class ImportActivitiesService {
       unitPrice,
       currency: SymbolProfile.currency,
       date: date.toString(),
-      symbol: SymbolProfile.symbol
+      symbol: SymbolProfile.symbol,
+      updateAccountBalance
     };
   }